feat: usb command submission
Yeah I know, this is done in an awful way in respect to me manually
building the commands. In the future, I want to make a library to build
PMBus commands and then use that library in both the microcontroller and
this application, which would deduplicate a ton of code that I dont
wanna write.